+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 - download progress bar
 - terminal user interface (TUI) for download configuration
 - automatically detect most recent object catalog for plotting
+- monochromatic cutout plotting
 
 ### Changed
 - integrate config validation into unionsdata edit and remove unionsdata validate